// next-app/src/common/interface.ts
export interface Brand {
  id: string;
  name: string;
  slug?: string;
  status: boolean;
  description?: string;
  description1?: string;
  description2?: string;
  description3?: string;
  image1?: string;
  image2?: string;
  image3?: string;
}